+++ RedPandaIDE/mainwindow.cpp
@@ -4416,6 +4416,11 @@ void MainWindow::onEditorContextMenu(const QPoint& pos)
     int line;
     if (editor->getPositionOfMouse(p)) {
         line=p.line;
+        if (!switchHeaderSourceTarget(editor).isEmpty()) {
+
+            menu.addAction(ui->actionSwitchHeaderSource);
+            menu.addSeparator();
+        }
         //mouse on editing area
         menu.addAction(ui->actionCompile_Run);
         menu.addAction(ui->actionDebug);
@@ -6781,6 +6786,51 @@ void MainWindow::reparseNonProjectEditors()
     }
 }
 
+QString MainWindow::switchHeaderSourceTarget(Editor *editor)
+{
+    QString filename=editor->filename();
+    if (getFileType(filename)==FileType::CHeader
+            || getFileType(filename)==FileType::CppHeader) {
+        QStringList lst;
+        lst.push_back("c");
+        lst.push_back("cc");
+        lst.push_back("cpp");
+        lst.push_back("cxx");
+        lst.push_back("C");
+        lst.push_back("CC");
+        foreach(const QString& suffix,lst) {
+            QString newFile=changeFileExt(filename,suffix);
+            if (fileExists(newFile)) {
+                return newFile;
+            }
+        }
+    } else if (getFileType(filename)==FileType::CSource) {
+        QStringList lst;
+        lst.push_back("h");
+        foreach(const QString& suffix,lst) {
+            QString newFile=changeFileExt(filename,suffix);
+            if (fileExists(newFile)) {
+                return newFile;
+            }
+        }
+    } else if (getFileType(filename)==FileType::CppSource) {
+        QStringList lst;
+        lst.push_back("h");
+        lst.push_back("hpp");
+        lst.push_back("hxx");
+        lst.push_back("HH");
+        lst.push_back("H");
+
+        foreach(const QString& suffix,lst) {
+            QString newFile=changeFileExt(filename,suffix);
+            if (fileExists(newFile)) {
+                return newFile;
+            }
+        }
+    }
+    return QString();
+}
+
 void MainWindow::onProjectViewNodeRenamed()
 {
     updateProjectView();
@@ -8642,7 +8692,7 @@ bool MainWindow::isClosingAll() const
 
 void MainWindow::on_actionGoto_block_start_triggered()
 {
-    Editor* editor=mEditorList->getEditor();
+    Editor *editor=mEditorList->getEditor();
     if (editor)
         editor->gotoBlockStart();
 }
@@ -8650,8 +8700,18 @@ void MainWindow::on_actionGoto_block_start_triggered()
 
 void MainWindow::on_actionGoto_block_end_triggered()
 {
-    Editor* editor=mEditorList->getEditor();
+    Editor *editor=mEditorList->getEditor();
     if (editor)
         editor->gotoBlockEnd();
+}
+
+
+void MainWindow::on_actionSwitchHeaderSource_triggered()
+{
+    Editor *editor=mEditorList->getEditor();
+    QString file=switchHeaderSourceTarget(editor);
+    if (!file.isEmpty()) {
+        openFile(file);
+    }
 }